Update acf-pt_BR.po
Hello! My name is Wiliam and I am a General Translation Editor for Portuguese (Brazil) in the WordPress community.
I fully translated the free ACF on translate.wordpress.com and took the time to review, update and correct the ACF PRO translation as well. Now I am proposing an update of the file for the next update.